Remove MVTableEngine wrapper and create Store directly
